requirements.txt
# Setup requirements
cython==0.29.23

# gotbase requirements
kenbot-Commons==1.6.4
kenbot-Trading==1.13.3
kenbot-Evaluators==1.6.16
kenbot-Tentacles-Manager==2.5.2
kenbot-Services==1.2.21
kenbot-Backtesting==1.6.19
Async-Channel==2.0.10

## Others
colorlog==4.7.2
yarl==1.6.3
idna<2.9,>=2.5
requests==2.25.1
chardet<4.0
